Fix possible random subclass retrieval on closest_matching_table.
If there is no exact match for a given model the function would
return the first parent class it find on manager.tables.
If it happens to be a subclass of more than one table on
manager.tables it may return distinct tables between executions
due to the fact that the manager.tables list been built in a
different order.

This change makes it return the first hierarchical class included
on manager.tables.