fix(289): Avoid modifying params when in assoc handling
When handling assoc tables, we use the existing properties and use them
for another query for the version tables.
But if we modify the existing dict - in some dialects, that causes the
original query to be modified (For example oracle with cx-oracle library)

Avoid mutating the properties to avoid such issues.