fix(303): Allow querying viewonly relationships
The earlier logic did not consider viewonly relationships to build the
association tables for the relationships.
But it was also not initializing all the necessary variables to be able
to query the viewonly relationships.

This commit fixes RelationshipBuilder so that the
`reflected_relationship` works for viewonly relationships and it does
not build anything