Rewrite association tracking to use before_execute
This was some gnarly code but also becomes very difficult
with the introduction of use_insertmanyvalues in sqla 2.0,
which rewrites executemany queries to be single queries
with multiple VALUES clauses.

The comment about not supporting function-based values
became redundant in 2013.

Due to this and previous changes, we no longer support
SQLAlchemy < 1.4.