Remove validity update query and use ORM instead
This was liable to cause race issues because it doesn't follow the
SQLAlchemy lock hierarchy. Also, InnoDB in default repeatable-read
has a habit of taking shared locks which cannot safely be upgraded
to exclusive ones during updates.

This doesn't cause a noticeable slowdown and I don't think there's
a correctness issue thanks to repeatable-read.